Move BoundsMap functionality into Map
As requested, there is no validation of bounds vs. center/zoom modes to set the view. Last updated wins.
