+++ lib/__tests__/Map.js
@@ -83,6 +83,16 @@ describe('Map', function () {
     expect(mapLeaflet.getZoom()).toBe(zoom);
   });
 
+  it('sets bounds', function () {
+    var bounds = [[0, 0], [2, 2]];
+    var component = _react2['default'].createElement(Map, { bounds: bounds });
+    var instance = (0, _reactDom.render)(component, document.getElementById('test'));
+    var mapLeaflet = instance.leafletElement;
+
+    expect(mapLeaflet.getCenter().lat).toBeCloseTo(1);
+    expect(mapLeaflet.getCenter().lng).toBeCloseTo(1);
+  });
+
   it('updates center and zoom props', function () {
     var Component = (function (_React$Component2) {
       _inherits(Component, _React$Component2);
@@ -131,5 +141,51 @@ describe('Map', function () {
     expect(mapLeaflet.getCenter().lat).toBe(2.3);
     expect(mapLeaflet.getCenter().lng).toBe(4.5);
     expect(mapLeaflet.getZoom()).toBe(12);
+  });
+
+  it('updates bounds props', function () {
+    var Component = (function (_React$Component3) {
+      _inherits(Component, _React$Component3);
+
+      function Component() {
+        _classCallCheck(this, Component);
+
+        _get(Object.getPrototypeOf(Component.prototype), 'constructor', this).call(this);
+        this.state = {
+          bounds: [[0, 0], [2, 2]]
+        };
+      }
+
+      _createClass(Component, [{
+        key: 'getLeafletMap',
+        value: function getLeafletMap() {
+          return this.refs.map.leafletElement;
+        }
+      }, {
+        key: 'updatePosition',
+        value: function updatePosition() {
+          this.setState({
+            bounds: [[0, 0], [-2, -2]]
+          });
+        }
+      }, {
+        key: 'render',
+        value: function render() {
+          return _react2['default'].createElement(Map, { bounds: this.state.bounds, ref: 'map' });
+        }
+      }]);
+
+      return Component;
+    })(_react2['default'].Component);
+
+    var instance = (0, _reactDom.render)(_react2['default'].createElement(Component, null), document.getElementById('test'));
+    var mapLeaflet = instance.getLeafletMap();
+
+    expect(mapLeaflet.getCenter().lat).toBeCloseTo(1);
+    expect(mapLeaflet.getCenter().lng).toBeCloseTo(1);
+
+    instance.updatePosition();
+    expect(mapLeaflet.getCenter().lat).toBeCloseTo(-1);
+    expect(mapLeaflet.getCenter().lng).toBeCloseTo(-1);
   });
 });
