+++ src/__tests__/Map.js
@@ -45,6 +45,16 @@ describe('Map', () => {
     expect(mapLeaflet.getZoom()).toBe(zoom);
   });
 
+  it('sets bounds', () => {
+    const bounds = [[0, 0], [2, 2]];
+    const component = <Map bounds={bounds} />;
+    const instance = render(component, document.getElementById('test'));
+    const mapLeaflet = instance.leafletElement;
+
+    expect(mapLeaflet.getCenter().lat).toBeCloseTo(1);
+    expect(mapLeaflet.getCenter().lng).toBeCloseTo(1);
+  });
+
   it('updates center and zoom props', () => {
     class Component extends React.Component {
       constructor() {
@@ -79,4 +89,36 @@ describe('Map', () => {
     expect(mapLeaflet.getCenter().lng).toBe(4.5);
     expect(mapLeaflet.getZoom()).toBe(12);
   });
+
+  it('updates bounds props', () => {
+    class Component extends React.Component {
+      constructor() {
+        super();
+        this.state = {
+          bounds: [[0, 0], [2, 2]],
+        };
+      }
+      getLeafletMap() {
+        return this.refs.map.leafletElement;
+      }
+      updatePosition() {
+        this.setState({
+          bounds: [[0, 0], [-2, -2]],
+        });
+      }
+      render() {
+        return <Map bounds={this.state.bounds} ref='map' />;
+      }
+    }
+    const instance = render(<Component />, document.getElementById('test'));
+    const mapLeaflet = instance.getLeafletMap();
+
+    expect(mapLeaflet.getCenter().lat).toBeCloseTo(1);
+    expect(mapLeaflet.getCenter().lng).toBeCloseTo(1);
+
+    instance.updatePosition();
+    expect(mapLeaflet.getCenter().lat).toBeCloseTo(-1);
+    expect(mapLeaflet.getCenter().lng).toBeCloseTo(-1);
+  });
+
 });
